+++ plugins/vaultgpt/scripts/vaultgpt/tests/__init__.py
@@ -1,0 +1,2 @@
+"""VaultGPT script tests."""
+